requirements.txt
requests
faker
holidays
python-dotenv